Moved authentication_trait into tests phpunit
